requirements.txt
accelerate
datasets
deepspeed>=0.15.0
einops
flash-attn>=2.7.0
jsonlines
loralib
numpy
openai
packaging
pandas
peft
pyyaml
requests
tensorboard
torch
tqdm
transformers>=4.46.0
wandb
